requirements.txt
torch
torch-geometric
scikit-learn
pandas
numpy
matplotlib
networkx
tqdm
transformers
streamlit
plotly
seaborn
pillow
